Add of the updater script
The script ```photon_update.sh``` in the ```update/``` folder is an
updater for the Photon's Elasticsearch indexes.

It can be used manually launching the script directly via CLI but also
in crontab.

We recommend to launch it once a week thrue a crontab job (the Photon's
Elasticsearche indexes are updated on [the source repository once a
week](https://github.com/komoot/photon#installation)).